backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
groq
langgraph
e2b
pytest
flake8
mypy
sse-starlette
gitpython
supabase
python-jose[cryptography]
psycopg2-binary
